Add product delete flow in API, service, repository and UI

+++ app/repositories/product_repository.py
@@ -1,6 +1,7 @@
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
+from app.exceptions.exceptions import DataNotFoundError
 from app.models.product_model import Product
 
 
@@ -18,8 +19,15 @@ class ProductRepository:
         stmt = select(Product).where(Product.id == product_id)
         product = self.db.execute(stmt).scalars().first()
         for key, value in data.items():
-            if value:
+            if value is not None:
                 setattr(product, key, value)
+        self.db.commit()
+        self.db.refresh(product)
+        return product
+
+    def create_product(self, data: dict):
+        product = Product(**data)
+        self.db.add(product)
         self.db.commit()
         self.db.refresh(product)
         return product
@@ -30,5 +38,15 @@ class ProductRepository:
         return products
 
     def delete(self, product: Product):
+        self.db.delete(product)
+        self.db.commit()
+
+    def delete_product(self, product_id: int):
+        stmt = select(Product).where(Product.id == product_id)
+        product = self.db.execute(stmt).scalars().first()
+
+        if product is None:
+            raise DataNotFoundError()
+
         self.db.delete(product)
         self.db.commit()
